docker-images/python/requirements.txt
imutils
paho-mqtt
numpy